+++ src/plone.server/CHANGELOG.rst
@@ -1,6 +1,11 @@
 1.0a9 (unreleased)
 ------------------
 
+- Provide base plone.server.commands.Command class to provide your own commands.
+  Commands have been moved in code so you'll need to re-run buildout to get
+  pserver to work after this update.
+  [vangheem]
+
 - Automatically give authenticated users new `plone.Authenticated` role
   [vangheem]